Add index.html and main.js to bundle
This change adds the web GUI to the project bundle.

I didn't think it would actually work, but it does! As long as the
script tag does not have a type="module" (because that would
trigger CORS trouble), this will let you open the web GUI directly
from CIRCUITPY/index.html.